refactoring: add dir with controllers
