+++ controllers/contacts/getAll.js
@@ -1,0 +1,13 @@
+const {listContacts} = require("../../models/contacts");
+
+const getAllContacts = async (req, res, next) => {
+  try {
+    const contacts = await listContacts();
+    res.status(200).json(contacts);
+  } catch (error) {
+    next(error);
+    // res.status(500).json({ message: "Server error!!!" });
+  }
+};
+
+module.exports = getAllContacts;
