+++ controllers/contacts/index.js
@@ -1,0 +1,7 @@
+const getAllContacts = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const updateById = require("./updateById");
+const removeById = require("./removeById");
+
+module.exports = { getAllContacts, getById, add, updateById, removeById };
