+++ controllers/index.js
@@ -1,0 +1,3 @@
+const contacts = require("./contacts");
+
+module.exports = { contacts };
