+++ models/contacts.json
@@ -48,21 +48,9 @@
     "phone": "(233) 738-2360"
   },
   {
-    "id": "e6ywwRe4jcqxXfCZOj_1e",
-    "name": "Thomas Lucas",
-    "email": "nec@Nulla.com",
-    "phone": "(704) 398-7993"
-  },
-  {
     "id": "rsKkOQUi80UsgVPCcLZZW",
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "name": "All Ray",
-    "email": "nulla.ante@vestibul.co.ff",
-    "phone": "(112) 222-2222",
-    "id": "1f2b136b-b7fa-4a13-8d4e-06746b48987a"
   }
 ]
